# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Astral Modulator Public API
  version: 2.0.0-scaffold
  license:
    name: MIT
    identifier: MIT
  description: |
    人类与多 LLM Agent 的协作控制面。CLI（astral-cli 仓库）与 Web GUI 共用本协议。

    v2（容器化任务树，TODO.md D15）：workspace 与 task 都是容器，子任务集合
    统一为 GET/POST <container>/children；全量平铺列表与 /tasks/search 已移除，
    平面查询走 /workspaces/{id}/task-search。protocol_version 升为 2，
    与 v1 不兼容（破坏性变更，开发期不做兼容窗口）。

    已裁决的文档分歧（详见 TODO.md「文档分歧裁决」D1-D4）：
    - Device flow 采用 `/auth/device/authorizations*` 路径。
    - 当前身份端点为 `/auth/me`。

    通用约定：
    - 所有时间戳为 RFC3339 UTC。
    - ID 为 `prefix_uuidv7` 形式的不透明字符串，客户端不得解析内部结构。
    - 错误响应统一使用 ErrorEnvelope；CLI 依赖稳定 error.code。
    - 乐观并发统一使用 body 内 expected_revision（protocol §6 方案 A）。
servers:
  - url: /api/v1
    description: 路径前缀固定 /api/v1（不随协议大版本变化）；协议大版本以
      /.well-known/astral 的 protocol_version 为准
...
security:
  - bearerAuth: []
paths:
...
  /workspaces:
    post:
      tags: [workspaces]
      summary: 创建 workspace（支持 Idempotency-Key）
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceCreate" }
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "409": { $ref: "#/components/responses/Error" }
    get:
      tags: [workspaces]
      summary: 列出可见 workspace；支持 ?name= 精确 name/slug 解析（astral init 依赖）
      operationId: listWorkspaces
      parameters:
        - { name: name, in: query, schema: { type: string }, description: exact name 或 slug }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: workspace 分页列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspacePage"
  /workspaces/{workspace_id}:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
    get:
      tags: [workspaces]
      summary: workspace 详情（init 最终绑定校验用）
      operationId: getWorkspace
      responses:
        "200":
          description: 详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [workspaces]
      summary: 改名（workspace_id 不变，改名不破坏 CLI 绑定）
      operationId: updateWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceUpdate" }
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
...
  /workspaces/{workspace_id}/approvals:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
    post:
      tags: [approvals]
      summary: 发起高风险动作审批（MVP：membership.promote_owner）
      operationId: createApproval
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApprovalCreate" }
      responses:
        "201":
          description: 已创建（status=requested）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Approval" }
    get:
      tags: [approvals]
      summary: 审批列表（可按 status 过滤）
      operationId: listApprovals
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [requested, approved, rejected, expired, executed] }
      responses:
        "200":
          description: 审批分页列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalPage" }
  /approvals/{approval_id}/approve:
    post:
      tags: [approvals]
      summary: 批准并在同事务执行（仅 workspace owner）
      operationId: approveApproval
      parameters:
        - { name: approval_id, in: path, required: true, schema: { $ref: "#/components/schemas/Id" } }
      responses:
        "200":
          description: 已批准并执行（status=executed）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Approval" }
        "409":
          $ref: "#/components/responses/Conflict"
  /approvals/{approval_id}/deny:
    post:
      tags: [approvals]
      summary: 拒绝（仅 workspace owner）
      operationId: denyApproval
      parameters:
        - { name: approval_id, in: path, required: true, schema: { $ref: "#/components/schemas/Id" } }
      responses:
        "200":
          description: 已拒绝（status=rejected）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Approval" }
        "409":
          $ref: "#/components/responses/Conflict"
...
  /workspaces/{workspace_id}/invitations:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
    post:
      tags: [workspaces]
      summary: 签发一次性邀请码（A5；需 human session + workspace:manage_members，agent credential 403）
      operationId: createInvitation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationCreate" }
      responses:
        "201":
          description: 已签发；code 明文仅本次返回，invite_url 为拼好的注册链接
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationCreated" }
        "400": { $ref: "#/components/responses/Error" }   # VALIDATION_FAILED（role/expires_in）
        "403": { $ref: "#/components/responses/Error" }   # agent credential / INSUFFICIENT_SCOPE
        "404": { $ref: "#/components/responses/Error" }   # 非成员不可见（WORKSPACE_NOT_FOUND）
    get:
      tags: [workspaces]
      summary: 邀请列表（授权同签发；不含 code——库中只有 hash）
      operationId: listInvitations
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [invited, redeemed, revoked] }
      responses:
        "200":
          description: 邀请分页列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationPage" }
        "400": { $ref: "#/components/responses/Error" }   # status 过滤值非法
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /invitations/{invitation_id}/revoke:
    post:
      tags: [workspaces]
      summary: 撤销邀请（invited→revoked；对已关闭邀请幂等 204；授权同签发）
      operationId: revokeInvitation
      parameters:
        - { name: invitation_id, in: path, required: true, schema: { $ref: "#/components/schemas/Id" } }
      responses:
        "204": { description: 已撤销（或邀请本已关闭，幂等） }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Human access token 或 Agent/Service credential（不共用）
    cookieAuth:
      type: apiKey
      in: cookie
      name: astral_session
      description: Web GUI HttpOnly+Secure+SameSite session（与 CLI Bearer 同 actor 不同 session）
  parameters:
    WorkspaceId:
      name: workspace_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string, description: 不透明游标，取自 next_cursor }
...
  responses:
    Error:
      description: 统一错误 envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    Conflict:
      description: 状态冲突（409：REVISION_CONFLICT / APPROVAL_EXPIRED /
        TASK_ALREADY_CLAIMED 等，细分语义见 error.code 与 error.details）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    Id:
      type: string
      pattern: "^[a-z]{2,3}_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      description: 不透明 ID：prefix_uuidv7。前缀 usr/agt/svc/srv/ws/tsk/tgp/tag/doc/cfl/prs/msg/evt/cred/ses/dev/req/aud/apv/inv。
      examples: ["ws_0192ab34-56cd-7ef8-9a01-234567890abc"]
    IdOrNull:
      oneOf:
        - $ref: "#/components/schemas/Id"
        - type: "null"
    DateTimeOrNull:
      type: [string, "null"]
      format: date-time
...
    ErrorCode:
      type: string
      enum:
        - AUTH_REQUIRED
        - TOKEN_EXPIRED
        - TOKEN_REVOKED
        - INSUFFICIENT_SCOPE
        - SERVER_NOT_FOUND
        - WORKSPACE_NOT_FOUND
        - WORKSPACE_ALREADY_BOUND
        - WORKSPACE_NAME_TAKEN
        - TASK_NOT_FOUND
        - TASK_ALREADY_CLAIMED
        - TASK_LEASE_EXPIRED
        - TAG_PROPOSAL_EXPIRED
        - APPROVAL_EXPIRED        # approval 过期或已裁决（409）
        - INVITE_INVALID          # 邀请码不存在/已兑换/已撤销/已过期统一此码（防探测，400）
        - EMAIL_TAKEN             # email 已被注册（409，仅邀请码验证通过后可达）
        - TAG_ALREADY_EXISTS
        - REVISION_CONFLICT
        - DOCUMENT_CONFLICT
        - RATE_LIMITED
        - CLIENT_VERSION_UNSUPPORTED
        - VALIDATION_FAILED
        - NOT_FOUND               # 未知路由 / 无专用码的次级资源不存在（HTTP 404）
        - AUTHORIZATION_PENDING   # device flow 轮询（A1/RFC 8628），HTTP 400 + retryable
        - SLOW_DOWN               # device flow 轮询过快（A1），HTTP 400 + retryable
        - INTERNAL_ERROR
        - NOT_IMPLEMENTED   # 仅脚手架阶段（HTTP 501），正式端点不得返回
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, retryable, request_id]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message: { type: string }
            retryable: { type: boolean }
            details: { type: object, additionalProperties: true }
            request_id: { type: string }
...
    WorkspaceCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 200 }
        slug: { type: string, maxLength: 200 }
    WorkspaceUpdate:
      type: object
      minProperties: 1
      properties:
        name: { type: string, maxLength: 200 }
        slug: { type: string, maxLength: 200 }
    Workspace:
      type: object
      required: [id, name, slug, created_at]
      properties:
        id: { $ref: "#/components/schemas/Id" }
        name: { type: string }
        slug: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    InvitationRole:
      type: string
      enum: [viewer, contributor, maintainer]
      description: 邀请可授角色（永不含 owner——晋升必须走 approval，D8）
    InvitationCreate:
      type: object
      required: [role]
      properties:
        role: { $ref: "#/components/schemas/InvitationRole" }
        expires_in:
          type: integer
          minimum: 1
          maximum: 2592000
          default: 604800
          description: 有效期（秒）；缺省 7 天，上限 30 天
    Invitation:
      type: object
      required: [id, workspace_id, role, status, created_by, created_at, expires_at]
      properties:
        id: { $ref: "#/components/schemas/Id" }
        workspace_id: { $ref: "#/components/schemas/Id" }
        role: { $ref: "#/components/schemas/InvitationRole" }
        status:
          type: string
          enum: [invited, redeemed, revoked]
          description: expired 为派生态（invited 且 now > expires_at），不落库
        created_by: { $ref: "#/components/schemas/Id" }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        redeemed_by: { $ref: "#/components/schemas/IdOrNull" }
        redeemed_at: { $ref: "#/components/schemas/DateTimeOrNull" }
    InvitationCreated:
      allOf:
        - $ref: "#/components/schemas/Invitation"
        - type: object
          required: [code, invite_url]
          properties:
            code:
              type: string
              description: 邀请码明文（XXXXX-XXXXX-XXXXX-XXXXX，Crockford base32，100 bit 熵），仅签发响应返回一次
            invite_url:
              type: string
              format: uri
              description: "{WebBaseURL}/register?code=<code>；基址回退链 WebBaseURL→PublicURL→请求 Host"
    InvitationPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/Invitation" } }
...
    ApprovalAction:
      type: string
      enum: [membership.promote_owner]
    ApprovalStatus:
      type: string
      enum: [requested, approved, rejected, expired, executed]
    ApprovalCreate:
      type: object
      required: [action, target_actor_id]
      properties:
        action: { $ref: "#/components/schemas/ApprovalAction" }
        target_actor_id: { $ref: "#/components/schemas/Id" }
    Approval:
      type: object
      required: [id, workspace_id, action, target_actor_id, status, requested_by, expires_at, created_at]
      properties:
        id: { $ref: "#/components/schemas/Id" }
        workspace_id: { $ref: "#/components/schemas/Id" }
        action: { $ref: "#/components/schemas/ApprovalAction" }
        target_actor_id: { $ref: "#/components/schemas/Id" }
        status: { $ref: "#/components/schemas/ApprovalStatus" }
        requested_by: { $ref: "#/components/schemas/Id" }
        decided_by: { $ref: "#/components/schemas/IdOrNull" }
        decided_at: { $ref: "#/components/schemas/DateTimeOrNull" }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    ApprovalPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/Approval" } }

    # ---- 分页 envelope（protocol §5：cursor，不暴露 offset 语义）----
    PageMeta:
      type: object
      required: [next_cursor]
      properties:
        next_cursor:
          type: [string, "null"]
          description: null 表示没有更多
    WorkspacePage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/Workspace" } }
...